feat: comprehensive codebase updates for production readiness
- Enhance cryptographic keychain with improved backend support
- Add robust framework detection and repository metadata APIs
- Implement comprehensive security scanning capabilities
- Extend system metrics database with advanced monitoring
- Improve service supervisor with enhanced metrics and auth
- Add new API endpoints for codegraph analysis and repository insights
- Update UI components with better alert widget testing

These changes strengthen the platform's security, monitoring, and analysis capabilities for production deployment.

+++ crates/adapteros-service-supervisor/src/supervisor.rs
@@ -238,7 +238,7 @@ impl ServiceSupervisor {
             self.config
                 .services
                 .get(service.id())
-                .map(|config| -(config.startup_order as i32)) // Negative for reverse order
+                .map(|config| -config.startup_order) // Negative for reverse order
                 .unwrap_or(-999)
         });
